Add owner key
This object is to the describe the owner association, if it is different
from the space name. For example our space "Chaosstuff" is run by the
local CCC, Chaos Computer Club Lëtzebuerg. So it would be good to
include this information in the spaceapi.

This might also be useful in the future, if an association runs multiple
hackerspaces and every space would have it's own spaceapi to show the
opening status. With this apps could link those spaces together for
example.